requirements.txt
Sphinx
mock
requests
attrdict
python-magic